+++ pkg/local_object_storage/shard/bench_test.go
@@ -12,6 +12,7 @@ import (
 	"github.com/nspcc-dev/neofs-sdk-go/object"
 	oid "github.com/nspcc-dev/neofs-sdk-go/object/id"
 	oidtest "github.com/nspcc-dev/neofs-sdk-go/object/id/test"
+	"github.com/nspcc-dev/neofs-sdk-go/user"
 	"github.com/stretchr/testify/require"
 )
 
@@ -152,11 +153,10 @@ func prepareObjects(b *testing.B, creat func(testing.TB) common.Storage, objSize
 	require.NoError(b, ptt.Init())
 	b.Cleanup(func() { _ = ptt.Close() })
 
-	obj := object.New()
+	obj := object.New(cid.ID{1, 2, 3}, user.ID{1, 2, 3})
 	data := make([]byte, objSize)
 	_, _ = rand.Read(data)
 	obj.SetID(oid.ID{1, 2, 3})
-	obj.SetContainerID(cid.ID{1, 2, 3})
 	obj.SetPayload(data)
 
 	rawData := obj.Marshal()